sys_mon_util/requirements.txt
psutil>=5.9.0
requests>=2.31.0
python-daemon>=3.0.1; platform_system != "Windows"
pywin32>=306; platform_system == "Windows"
schedule>=1.2.0
pyyaml>=6.0.1
cryptography>=41.0.0
